+++ packages/bridge-ui/src/relayer-api/RelayerAPIService.ts
@@ -11,9 +11,10 @@ import { chainIdToTokenVaultAddress } from '../store/bridge';
 import { get } from 'svelte/store';
 import type { RelayerAPI, RelayerBlockInfo } from 'src/domain/relayerApi';
 
-class RelayerAPIService implements RelayerAPI {
+export class RelayerAPIService implements RelayerAPI {
   private readonly providerMap: Map<number, ethers.providers.JsonRpcProvider>;
   private readonly baseUrl: string;
+
   constructor(
     providerMap: Map<number, ethers.providers.JsonRpcProvider>,
     baseUrl: string,
@@ -194,5 +195,3 @@ class RelayerAPIService implements RelayerAPI {
     return blockInfoMap;
   }
 }
-
-export default RelayerAPIService;
